+++ README.md
@@ -17,3 +17,5 @@ osu!の譜面フォルダーから音楽ファイルをコピーします。
 - [ ] カバーアートもコピー・合成を可能に
 - [x] オプションを追加
 - [ ] 進捗状況を見られるように
+- [ ] pathlibに移行
+- [ ] TkInterDnD2の設定
